+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 - **Change:** Stop building readme when building add-on (#83)
 - **Fix:** Suppress DOM errors when attempting to convert markdown to HTML (#81)
 - **Fix:** Fix potential server error when attempting to copy files
+- **Fix:** Fixed an error with building the list of requirements if the version was `*` instead of an array
 
 ## 1.3.0 (`1030070`)