+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - **New:** Arguments must now be passed to `tck-devtools:build-readme` in order to build any of the 3 different output formats
 - **New:** A new argument can be passed to `tck-devtools:build-readme` in order to copy the resulting file to the `_no_upload` directory
+- **New:** Support for Xon's `require-soft` recommendations parameter in `addon.json`
 - **Change:** Stop building readme when building add-on (#83)
 - **Fix:** Suppress DOM errors when attempting to convert markdown to HTML (#81)
 - **Fix:** Fix potential server error when attempting to copy files