+++ CHANGELOG.md
@@ -6,9 +6,11 @@ CHANGELOG
 - **New:** Arguments must now be passed to `tck-devtools:build-readme` in order to build any of the 3 different output formats
 - **New:** A new argument can be passed to `tck-devtools:build-readme` in order to copy the resulting file to the `_no_upload` directory
 - **Change:** Stop building readme when building add-on (#83)
+- **Change:** Changed the wording on the "Add more phrase" button to "Add another phrase" (#82)
 - **Fix:** Suppress DOM errors when attempting to convert markdown to HTML (#81)
 - **Fix:** Fix potential server error when attempting to copy files
 - **Fix:** Fixed an error with building the list of requirements if the version was `*` instead of an array
+- **Fix:** Fixed being unable to save phrases if you accidentally clicked the "Add more phrase" button (#82)
 
 ## 1.3.0 (`1030070`)